Add files via upload
Business Understanding
Passenger satisfaction is of paramount importance in the airline business. Passenger satisfaction 
ensures Customer Loyalty, Positive Word of Mouth, Increased Revenue, Reduced Complaints 
and Disputes, Sustainable Growth etc. It is a critical component of a successful airline business. 
Airlines that prioritize and consistently deliver high levels of passenger satisfaction can enjoy a 
wide range of benefits, from increased revenue and customer loyalty to a strong brand image and
competitive advantage. There is an urgent need to understand the factor leading to dissatisfaction 
of the passenger